buck2: superconsole: take `self` instead of `&mut self` in finalize
Summary:
I'd like to make it possible for outputs to know when no further output is to
be expected, so to do this we need to prevent finalizing more than once.

In Buck2 we never actually call this more than once, and on the whole it seems
like the reasonable thing to do.

Reviewed By: ndmitchell, bobyangyf

Differential Revision: D34818037

fbshipit-source-id: fc0d8bd154ac13ebb9c5530e6bd65086583c7510