buck2: superconsole: non-blocking output
Summary:
Like it says in the title. This introduces a non-blocking output in the
Superconsole. The goal of this output is to prevent re-renders if stderr is not
accepting writes.

Visually, if someone is looking at the terminal ,the output will be fully
unchanged (because all writes will go through), whereas if nobody is looking
for a while, all intermediary renders will be skipped, but finalize (or, if
someone reconnects, the most recent render) will still produce the exact same
terminal view it would have produced if someone was in fact looking.

There are still cases where we'll blocked, namely in clear() and finalize(). We
don't use clear() in Buck2, so I've left it as-is, but we could support it
later by treating it the same way we do emit() (that is, don't do it, just set
a field to indicate that the next render should clear).

Note that the approach here _will_ drop frames if stderr is flushing very slowly,
this is intentional back-pressure: it's not useful to generate faster than the output can
flush them.

The next diff adds support for enabling this.

Reviewed By: ndmitchell

Differential Revision: D34818038

fbshipit-source-id: 6b207c6bec37e2cf49d2670b5a2e104b4845b679