buck2: superconsole: guarantee that errors are returned before finalize()
Summary:
Like it says in the title; This adds one more guarantee to the non-blocking
output, which is that errors will be returned in calls to `output()` and
`finalize()` (whereas until now you'd usually get it in `output()` and it was
guaranteed in `finalize()`).

This also makes it possible to continue using the output after it it's hit an
error. While that's not very useful (if you're writing to stderr then an error
most likely means EPIPE and that will continue giving errors forever), it is
true of the blocking output so it doesn't hurt to be consistent.

I also added a little documentation explaining the guarnatees this output does
make.

Reviewed By: ndmitchell

Differential Revision: D34865534

fbshipit-source-id: 6ad403d9eaa18756c4f0146eb00e74035b81db41