Fix an off-by-one bug in constrained consoles
Summary: If the console is 5 lines high, we can only emit 4 lines, because we like to leave a blank line at the bottom where the cursor lives. It would be possible to restructure superconsole to leave the cursor at the bottom right of the final line, but until that is done, we have to account for that final line.

Reviewed By: bobyangyf

Differential Revision: D34936180

fbshipit-source-id: 6871e37daaa53a5eb48a36505c4c8cd9b63bc172