Treat empty terminal size as fall back too
Summary:
If the user has zero visible terminal, use the default size. Required because our CircleCI tests report they have 0x0 size, and if we actually use that, none of our tests work as they are sized too small.

I wonder if we actually need override_size and default_size as separate concepts, or if default_size is redundant with override. But this is the smallest change to make things work for now. I have a feeling though that running these tests in a small enough terminal window would fail today.

Reviewed By: stepancheg

Differential Revision: D39263327

fbshipit-source-id: f1bdd68e86d7ba76c58eb263603c14d9874e2ce7