Remove Expanding
Summary:
`Expanding` is stateful (which is already not optimal because it is very unclear that a component has hidden state).

In gets in the way of removing state (D44821871): when components are created on the fly, state cannot be shared. So this component might be misleading.

Also it contradicts documentation which states components are stateless:

https://www.internalfb.com/code/fbsource/[eb1e0d96a8da9fe7d89b75572f141f63287a17f4]/fbcode/buck2/superconsole/src/lib.rs?lines=22

Differential Revision: D44821867

fbshipit-source-id: d4a96194f0e50c0632712b965cbe1e9b1eb2600e