Remove State
Summary:
[Workplace post](https://fb.workplace.com/groups/buck2dev/posts/3436649489956432)

`State` is not statically typed, so it's hard to use it correctly.

For example, when some component is rendered conditionally (like turned on with a flag), we can forget (or accidentally remove) an item from `State`, and not notice it for long time.

Or for example, we can place in component something which is never used, and there's no easy way to check it is not really needed.

Previous diffs in the stack removed use of `Stack`, this diff only removes the type, and removes it from function signatures.

Differential Revision: D44821871

fbshipit-source-id: 45b05316130bdc101ed7a57fa377c6e17660499e