Add raw Span construction for emit/emit_aux with unrestricted whitespace
Summary:
## Context
https://fb.workplace.com/groups/buck2users/posts/3984509478472070/?comment_id=3984679978455020&reply_comment_id=3984714815118203

In the post, user want to output `\t` in output, but because right now all the output is will go though superconsole, superconsole will remove all the `whitespaces` except spaces, which case the issue.

The purpose of this is to make sure the precise characters len when rendering the canvas because it has alignment. But in `emit/emit_aux` it will only output above the canvas, without need for the alignments.

The goal is to ensure precise character length alignment when rendering the canvas. However, in the `emit/emit_aux` aces, output is placed above the canvas without requiring alignment.

## This diff

This diff add  raw Span construction for emit/emit_aux with unrestricted whitespace. And use this for emitting messages in buck2.

Reviewed By: podtserkovskiy

Differential Revision: D75811089

fbshipit-source-id: 00726d8db02cb9edab8312f194dcf4bd7860bfb6